+++ aws-classic-py-aws-load-balancer-controller-helm-release/__main__.py
@@ -4,19 +4,22 @@ import pulumi_eks as eks
 from pulumi import export, Output, ResourceOptions, get_stack, get_project
 import iam
 import vpc
+import pulumi_kubernetes as k8s
 from pulumi_kubernetes.core.v1 import Namespace
 from pulumi_kubernetes.helm.v3 import Release, ReleaseArgs, RepositoryOptsArgs
+from pulumi_kubernetes import Provider
 
+name = "demo"
 # create an iam role for eks
-role0 = iam.create_role("demo-py-role0")
+role0 = iam.create_role(f'{name}-eks-role0')
 
 # generic tags across cluster
 mytags ={"project_name":get_project(), "stack_name":get_stack()}
 
 # Create an EKS cluster.
-mycluster = eks.Cluster("demo-py-eks",
+mycluster = eks.Cluster(f'{name}-eks',
             skip_default_node_group = True,
-            instance_type = "t3a.medium",
+            instance_type = "t3a.small",
             version = "1.21",
             node_root_volume_size = 10,
             instance_roles=[role0],
@@ -32,7 +35,7 @@ mycluster = eks.Cluster("demo-py-eks",
             )
 
 # spot managed node group to save cost
-managed_nodegroup_spot_0 = eks.ManagedNodeGroup("demo-py-managed-nodegroup-spot-ng0",
+managed_nodegroup_spot_0 = eks.ManagedNodeGroup(f'{name}-managed-nodegroup-spot-ng0',
    cluster=mycluster.core, # TODO[pulumi/pulumi-eks#483]: Pass cluster directly.
    capacity_type = "SPOT",
    instance_types=["t3a.medium"],
@@ -45,9 +48,9 @@ managed_nodegroup_spot_0 = eks.ManagedNodeGroup("demo-py-managed-nodegroup-spot-
    opts=ResourceOptions(depends_on=[mycluster])
    )
 
+k8s_provider = Provider("mycluster_provider",kubeconfig=mycluster.kubeconfig)
 # Creatinga namespace
-namespace = Namespace("awslbcontroller-ns",
-                      opts=ResourceOptions(depends_on=[mycluster], provider=mycluster.provider))
+awslbcontroller_namespace = Namespace("awslb-controller-ns",opts=ResourceOptions(provider=k8s_provider))
 
 # Helm Release: https://www.pulumi.com/registry/packages/kubernetes/api-docs/helm/v3/release/
 release_args = ReleaseArgs(
@@ -55,18 +58,23 @@ release_args = ReleaseArgs(
     repository_opts=RepositoryOptsArgs(
         repo="https://aws.github.io/eks-charts"
     ),
-    version="1.3.3",
-    namespace=namespace.metadata["name"],
+    version="1.4.0",
+    namespace=awslbcontroller_namespace.metadata["name"],
     values={
         "clusterName": mycluster.core,
+        "createIngressClassResource": "true",
+        "ingressClassParams": {
+            "name": "alb-ingress-class-params"
         },
-    # By default Release resource will wait till all created resources
-    # are available. Set this to true to skip waiting on resources being
-    # available.
+        },
+#    # By default Release resource will wait till all created resources
+#    # are available. Set this to true to skip waiting on resources being
+#    # available.
     skip_await=False)   
-
-release = Release("aws-load-balancer-controller", args=release_args, opts=ResourceOptions(parent=namespace, provider=mycluster.provider))
+#
+release = Release("aws-load-balancer-controller", args=release_args, opts=ResourceOptions(provider=k8s_provider))
 status = release.status
+
 
 export("vpcname", vpc.vpc.id)
 export("cluster_name", mycluster.core.cluster.name)
@@ -74,6 +82,6 @@ export("managed_nodegroup_name", managed_nodegroup_spot_0.node_group.node_group_
 export("managed_nodegroup_capacity_type", managed_nodegroup_spot_0.node_group.capacity_type)
 export("managed_nodegroup_version", managed_nodegroup_spot_0.node_group.version)
 export("kubeconfig", Output.secret(mycluster.kubeconfig))
-export("k8s_namespace", namespace.id)
+export("k8s_namespace", awslbcontroller_namespace.id)
 export("k8s_chart",status["chart"])
 export("k8s_chart_app_version",status["app_version"])